Move user helper methods to separate file
